feat: Adding audit fields to product

+++ domain/src/main/java/postech/soat/tech/challenge/model/Product.java
@@ -3,6 +3,8 @@ package postech.soat.tech.challenge.model;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @AllArgsConstructor
 @Getter
 public class Product {
@@ -10,8 +12,8 @@ public class Product {
     private Long id;
     private String name;
     private String description;
-    private double price;
-    private int quantity;
+    private BigDecimal price;
+    private Integer quantity;
     private Category category;
 
 }
