+++ infrastructure/src/main/java/postech/soat/tech/challenge/api/request/ProductDTO.java
@@ -5,6 +5,7 @@ import lombok.Data;
 import postech.soat.tech.challenge.model.Category;
 import postech.soat.tech.challenge.model.Product;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -13,8 +14,8 @@ public class ProductDTO {
     private Long id;
     private String name;
     private String description;
-    private double price;
-    private int quantity;
+    private BigDecimal price;
+    private Integer quantity;
     private String category;
 
     public static Product toNewProduct(ProductDTO productDTO) {
